チャットオーバーレイの初期起動とチャット画面のUI調整

+++ JTSA/MainWindow.xaml.cs
@@ -186,12 +186,12 @@ namespace JTSA
 			}
 
             DAO_Setting.InsertUpdate(
-				(int)DAO_Setting.SettingName.RefreshToken,
+				DAO_Setting.SettingName.RefreshToken,
                 accessTokenResponse.refreshToken
 			);
 
             DAO_Setting.InsertUpdate(
-				(int)DAO_Setting.SettingName.ExpiresIn,
+				DAO_Setting.SettingName.ExpiresIn,
                 accessTokenResponse.expiresIn.ToString()
 			);
 
@@ -270,17 +270,17 @@ namespace JTSA
 
 			// --- 設定情報保存処理 ---
 			DAO_Setting.InsertUpdate(
-				(int)DAO_Setting.SettingName.UserName,
+				DAO_Setting.SettingName.UserName,
 				JTSAHelper.LoginName
 			);
 
 			DAO_Setting.InsertUpdate(
-				(int)DAO_Setting.SettingName.RefreshToken,
+				DAO_Setting.SettingName.RefreshToken,
 				accessTokenResponse.refreshToken
 			);
 
 			DAO_Setting.InsertUpdate(
-				(int)DAO_Setting.SettingName.ExpiresIn,
+				DAO_Setting.SettingName.ExpiresIn,
 				accessTokenResponse.expiresIn.ToString()
 			);
